test updated

+++ src/main/java/com/SpringBoot/SpringBootPart1/model/User.java
@@ -1,11 +1,11 @@
 package com.SpringBoot.SpringBootPart1.model;
 public class User {
 
-    public int userId;
+    public long userId;
     public String userName;
     public String userSurname;
 
-    public User(int userId, String userName, String userSurname){
+    public User(long userId, String userName, String userSurname){
         this.userId = userId;
         this.userName = userName;
         this.userSurname = userSurname;
@@ -16,4 +16,5 @@ public class User {
     public static String getUserSurname(String userSurname){
         return userSurname;
     }
+
 }
